Se hizo un refactoring en base al concept de las transactions, en el cual se ingreso un enumerador para poder lograr una mejor composicion y orden de los mismos, ademas de cambios verios.

+++ BilleteraVirtualSofttekBack/Services/AccountService.cs
@@ -129,6 +129,12 @@ namespace BilleteraVirtualSofttekBack.Services
         /// <inheritdoc/>
         public async Task<bool> DeleteAccountAsync(int id)
         {
+            var account = await _unitOfWork.AccountRepository.GetByIdAsync(id);
+
+            if(account.Balance > 1)
+            {
+                return false;
+            }
 
             var flag = await _unitOfWork.AccountRepository.Delete(id);
